fix n+1 for course, add top lecturer

+++ ELearning/src/main/java/com/dxhh/elearning/dto/response/CourseDetailsResponse.java
@@ -20,8 +20,8 @@ public class CourseDetailsResponse implements Serializable {
     private String description;
     private String background;
     private Double price;
-//    private List sections;
-    private Long countRegistration;
+    //    private List sections;
+    private Integer countRegistration;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy HH:mm:ss")
     private LocalDateTime createdDate;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy HH:mm:ss")
